fix: split freeze tests
