+++ gelidum/tests/gelidum_tests/test_freeze_objects.py
@@ -11,247 +11,15 @@ import unittest
 import warnings
 from typing import Dict, List, Union, Any
 from unittest.mock import patch
+
 from gelidum import FrozenException
 from gelidum import freeze
-from gelidum.dependencies import NUMPY_INSTALLED
-from gelidum.collections import frozendict, frozenlist, frozenzet
 from gelidum.frozen import FrozenBase, get_frozen_classes, clear_frozen_classes
 
 
-class TestFreeze(unittest.TestCase):
+class TestFreezeObjects(unittest.TestCase):
     def setUp(self) -> None:
         clear_frozen_classes()
-
-    def test_freeze_builtins(self):
-        self.assertEqual(3, freeze(3))
-        self.assertEqual(3.9, freeze(3.9))
-        self.assertEqual(True, freeze(True))
-        self.assertEqual(False, freeze(False))
-        self.assertEqual(None, freeze(None))
-        self.assertEqual(complex(1, 2), freeze(complex(1, 2)))
-        self.assertEqual(b"Bytes", freeze(b"Bytes"))
-        self.assertEqual("String", freeze("String"))
-
-    def test_freeze_bytearray(self):
-        self.assertEqual(b"Byte array", freeze(bytearray(b"Byte array")))
-
-    def test_freeze_dict(self):
-        frozen_obj: frozendict = freeze({"one": 1, "two": 2})
-
-        with self.assertRaises(FrozenException) as context_assignment:
-            frozen_obj["one"] = "another value"  # noqa
-
-        with self.assertRaises(FrozenException) as context_clear:
-            frozen_obj.clear()  # noqa
-
-        with self.assertRaises(FrozenException) as context_update:
-            frozen_obj.update({"three": 3})  # noqa
-
-        with self.assertRaises(FrozenException) as context_deletion:
-            del frozen_obj["one"]  # noqa
-
-        self.assertEqual(2, len(frozen_obj))
-        self.assertTrue("one" in frozen_obj)
-        self.assertTrue("two" in frozen_obj)
-        self.assertEqual(1, frozen_obj["one"])
-        self.assertEqual(2, frozen_obj["two"])
-        self.assertEqual("'frozendict' object is immutable", str(context_assignment.exception))
-        self.assertEqual("'frozendict' object is immutable", str(context_clear.exception))
-        self.assertEqual("'frozendict' object is immutable", str(context_update.exception))
-        self.assertEqual("'frozendict' object is immutable", str(context_deletion.exception))
-
-    def test_freeze_list(self):
-        frozen_list = freeze(["one", 2, "three", ["a", "b", "c", 4, 5]])
-        self.assertTrue(isinstance(frozen_list, frozenlist))
-        self.assertEqual(4, len(frozen_list))
-        self.assertEqual("one", frozen_list[0])
-        self.assertEqual(2, frozen_list[1])
-        self.assertEqual("three", frozen_list[2])
-        self.assertTrue(isinstance(frozen_list[3], frozenlist))
-        self.assertEqual(5, len(frozen_list[3]))
-        self.assertEqual("a", frozen_list[3][0])
-        self.assertEqual("b", frozen_list[3][1])
-        self.assertEqual("c", frozen_list[3][2])
-        self.assertEqual(4, frozen_list[3][3])
-        self.assertEqual(5, frozen_list[3][4])
-
-    def test_freeze_list_inplace_true_deprecated_parameter(self):
-        with warnings.catch_warnings(record=True) as caught_warnings:
-            frozen_list = freeze(["one", 2, "three"], inplace=True)
-
-        self.assertTrue(isinstance(frozen_list, frozenlist))
-        self.assertEqual(3, len(frozen_list))
-        self.assertEqual("one", frozen_list[0])
-        self.assertEqual(2, frozen_list[1])
-        self.assertEqual("three", frozen_list[2])
-        self.assertEqual(1, len(caught_warnings))
-        self.assertEqual(
-            "Use of inplace is deprecated and will be removed in next major version (0.6.0)",
-            str(caught_warnings[0].message),
-        )
-
-    def test_freeze_tuple(self):
-        self.assertEqual(("one", 2, "three"), freeze(("one", 2, "three")))
-
-    def test_freeze_set(self):
-        self.assertEqual(frozenzet(["one", 2, "three"]), freeze({"one", 2, "three"}))
-
-    def test_freeze_simple_dataclass(self):
-        from dataclasses import dataclass
-
-        @dataclass
-        class Dummy:
-            attr1: str
-            attr2: str
-            attr3: str = "0"
-
-        dummy = Dummy(attr1="1", attr2="2", attr3="3")
-        frozen_dummy_on_freeze_copy = freeze(dummy, on_freeze="copy")
-        frozen_dummy_inplace = freeze(dummy, on_freeze="inplace")
-
-        with self.assertRaises(FrozenException) as context_on_freeze_copy:
-            frozen_dummy_on_freeze_copy.attr1 = "2"
-
-        with self.assertRaises(FrozenException) as context_inplace:
-            frozen_dummy_inplace.attr2 = "2"
-
-        self.assertEqual("Can't assign attribute 'attr1' on immutable instance", str(context_on_freeze_copy.exception))
-        self.assertEqual("Can't assign attribute 'attr2' on immutable instance", str(context_inplace.exception))
-        self.assertEqual(id(dummy), id(frozen_dummy_inplace))
-        self.assertNotEqual(id(dummy), id(frozen_dummy_on_freeze_copy))
-
-    def test_freeze_nested_dataclass(self):
-        from dataclasses import dataclass
-
-        @dataclass
-        class ItemType:
-            id: int
-            name: str
-
-        @dataclass
-        class Item:
-            id: int
-            name: str
-            order: int
-            type: ItemType
-
-        @dataclass
-        class Database:
-            items: List[Item]
-
-        item_type1 = ItemType(1, "item type 1")
-        item_type2 = ItemType(2, "item type 2")
-
-        dummy_database = Database(
-            items=[
-                Item(1, "item 1", 5, item_type1),
-                Item(2, "item 2", 1, item_type1),
-                Item(3, "item 3", 99, item_type2),
-            ]
-        )
-        frozen_dummy_on_freeze_copy = freeze(dummy_database, on_freeze="copy")
-        frozen_dummy_inplace = freeze(dummy_database, on_freeze="inplace")
-
-        with self.assertRaises(FrozenException) as context_on_freeze_copy_items_assignment:
-            frozen_dummy_on_freeze_copy.items = []
-
-        with self.assertRaises(FrozenException) as context_on_freeze_copy_item_assignment:
-            frozen_dummy_on_freeze_copy.items[0] = Item(9, "item 9", 99, item_type1)
-
-        with self.assertRaises(FrozenException) as context_inplace:
-            frozen_dummy_inplace.attr2 = "2"
-
-        self.assertEqual(
-            "Can't assign attribute 'items' on immutable instance",
-            str(context_on_freeze_copy_items_assignment.exception),
-        )
-        self.assertEqual("'frozenlist' object is immutable", str(context_on_freeze_copy_item_assignment.exception))
-        self.assertEqual("Can't assign attribute 'attr2' on immutable instance", str(context_inplace.exception))
-        self.assertEqual(id(dummy_database), id(frozen_dummy_inplace))
-        self.assertNotEqual(id(dummy_database), id(frozen_dummy_on_freeze_copy))
-
-    def test_freeze_function_and_reading_attributes(self):
-        def times(x: int) -> int:
-            return x * times.factor
-
-        times.factor = 3
-
-        frozen_times = freeze(times)
-
-        self.assertEqual(3, frozen_times.factor)
-        self.assertNotEqual(id(times), id(frozen_times))
-
-    def test_freeze_function_inplace_is_ignored(self):
-        def times(x: int) -> int:
-            return x * times.factor
-
-        times.factor = 3
-
-        frozen_times = freeze(times, on_freeze="inplace")
-
-        self.assertEqual(3, frozen_times.factor)
-        self.assertNotEqual(id(times), id(frozen_times))
-
-    def test_freeze_function_and_writing_attributes_with_exception(self):
-        def times(x: int) -> int:
-            return x * times.factor
-
-        times.factor = 3
-
-        frozen_times = freeze(times)
-
-        with self.assertRaises(FrozenException) as context:
-            frozen_times.factor = 10
-
-        self.assertEqual("Can't assign attribute 'factor' on immutable instance", str(context.exception))
-
-    def test_freeze_function_and_writing_frozen_attributes_with_exception(self):
-        def times(x: int) -> int:
-            if x not in times.cache:
-                times.cache[x] = "some kind of slow and costly process"
-            return times.cache[x]
-
-        times.cache = {}
-
-        frozen_times = freeze(times)
-
-        with self.assertRaises(FrozenException) as context:
-            frozen_times.cache[10] = 789
-
-        self.assertEqual("'frozendict' object is immutable", str(context.exception))
-
-    def test_freeze_function_and_writing_attributes_with_function(self):
-        def times(x: int) -> int:
-            return x * times.factor
-
-        times.factor = 3
-
-        frozen_times = freeze(times, on_update="warning")
-
-        with warnings.catch_warnings(record=True) as caught_warnings:
-            frozen_times.factor = 10
-
-        self.assertListEqual(
-            [
-                "Can't assign attribute 'factor' on immutable instance"
-            ],
-            [str(warn.message) for warn in caught_warnings],
-        )
-
-
-    def test_freeze_objects_with_class_constants(self):
-        class Dummy(object):
-            CONSTANT1 = 1
-            CONSTANT2 = 2
-
-            def __init__(self, value: int):
-                self.attr = value
-
-        dummy = Dummy(1)
-        frozen_dummy = freeze(dummy, on_freeze="copy", save_original_on_copy=False)
-
-        self.assertEqual(1, frozen_dummy.CONSTANT1)
-        self.assertEqual(2, frozen_dummy.CONSTANT2)
 
     def test_freeze_objects_with__call__method(self):
         class Dummy(object):
@@ -1348,24 +1116,3 @@ class TestFreeze(unittest.TestCase):
         self.assertEqual(dummy1, my_dict[frozen_dummy1])
         self.assertTrue(frozen_dummy2 in my_dict)
         self.assertEqual(dummy2, my_dict[frozen_dummy2])
-
-    @unittest.skipUnless(NUMPY_INSTALLED, "numpy is not installed, TestFreeze.test_freeze_ndarray test skipped")
-    def test_freeze_ndarray(self):
-        import numpy as np
-
-        array = np.array([1, 2, 3])
-        frozen_array = freeze(array, on_freeze="copy")
-        frozen_array_copy = frozen_array.copy()
-
-        self.assertEqual(array.shape, frozen_array.shape)
-        self.assertEqual(array.shape, frozen_array_copy.shape)
-        self.assertEqual(array[0], frozen_array[0])
-        self.assertEqual(array[1], frozen_array[1])
-        self.assertEqual(array[2], frozen_array[2])
-        self.assertEqual(array[0], frozen_array_copy[0])
-        self.assertEqual(array[1], frozen_array_copy[1])
-        self.assertEqual(array[2], frozen_array_copy[2])
-        self.assertTrue(isinstance(frozen_array, FrozenBase))
-        self.assertTrue(isinstance(frozen_array, np.ndarray))
-        self.assertTrue(isinstance(frozen_array_copy, FrozenBase))
-        self.assertTrue(isinstance(frozen_array_copy, np.ndarray))
